Extract output verification to common file
